删除不必要的source操作 http://lean.xiaojukeji.com/project/ULTRON_REQ-1030/board

+++ src/main/bin/jar_generator.sh
@@ -1,16 +1,4 @@
 #!/bin/bash
-if [[ -f /etc/profile ]]; then
-    . /etc/profile
-fi
-
-if [[ -f ~/.bash_profile ]]; then
-    . ~/.bash_profile
-fi
-
-if [[ $3 == "" || $3 != *.jar ]]; then
-    echo "输入正确的jar路径，确保以.jar结尾"
-    exit 1
-fi
 
 function gen_classpath(){
     for element in `ls $LIB_DIR`
@@ -23,11 +11,16 @@ function gen_classpath(){
 declare -r HOME_DIR=$(cd $(dirname $0); cd ..; pwd)
 declare -r LIB_DIR="${HOME_DIR}/lib/thrift/$1"
 
+if [[ $3 == "" || $3 != *.jar ]]; then
+    echo "输入正确的jar路径，确保以.jar结尾"
+    exit 1
+fi
+
 if [[ -d classdir ]]; then
     rm -rf classdir
 fi
-
 mkdir classdir
+
 gen_classpath $1
 
 echo "version   -> $1"
